Mex upgrades are now never paused for throttle reasons.
I think that was causing the bug where, if you are low on energy, the
auto-mex-upgrade-and-pause-feature would start a mex upgrade as usual
but the pause would be registered as an energy-throttle pause, not a
auto-mex-upgrade-and-pause-feature pause. That would then result
in the upgrading t1 mex unintentionally getting unpaused once you had
more energy, which can obviously be disastraous.