Restore compatability with newest dev branch
On the current dev branch the LOC() function, which gets called in the
UIUtil.CreateText() function got changed to only accept string, removing
the implicit type casting from other types to strings that was there
before.
As such, the second argument of UIUtil.CreateText() now must be a string
and not an int as it was before.
While the backwards compatibilty of LOC with it's implicit type casting
may be introduced in a later commit, we may as well also just change the
arguments of UIUtil.CreateText() here as calling it with a string
instead of an int as label is less sloppy, i.e. better, anyway.