Fix bug where freshly built mexes would upgrade on their own
This happened when the auto-uprade-and-then-pause feature was turned on,
but due to the bug, only the upgrade part of that worked for some mexes,
not the pausing.
This was because the unit IDs used to keep track of mexes and
everything else that's pausable reuses ID numbers when they are
available. They become available when e.g. a paused mex dies, but the
state table the mod uses doesn't always get cleared of old, already dead
units. As a result, a new unit can be given an old ID which
sometimes already has a state table entry, telling the mod it's
suppposed to be paused, even if it isn't, resulting in the mod just
skipping the unit.
We fix this here not by fixing the state table, which would have been
the correct way to resolve it, but by just giving possibly redundant
pause commands to units, that may or may not already be paused to make
sure that they actually are.